+++ src/McOperator/Builders/ConfigMapBuilder.cs
@@ -105,7 +105,7 @@ public static class ConfigMapBuilder
             ["app.kubernetes.io/name"] = "minecraft-server",
             ["app.kubernetes.io/instance"] = server.Name(),
             ["app.kubernetes.io/managed-by"] = "mc-operator",
-            ["minecraft.dhv.sh/server-name"] = server.Name(),
-            ["minecraft.dhv.sh/config"] = "true",
+            ["mc-operator.dhv.sh/server-name"] = server.Name(),
+            ["mc-operator.dhv.sh/config"] = "true",
         };
 }