+++ src/McOperator/Builders/StatefulSetBuilder.cs
@@ -307,6 +307,6 @@ public static class StatefulSetBuilder
             ["app.kubernetes.io/name"] = "minecraft-server",
             ["app.kubernetes.io/instance"] = server.Name(),
             ["app.kubernetes.io/managed-by"] = "mc-operator",
-            ["minecraft.dhv.sh/server-name"] = server.Name(),
+            ["mc-operator.dhv.sh/server-name"] = server.Name(),
         };
 }